(wip) reorganize tests to mirror source files
