+++ kimera_pgmo/include/kimera_pgmo/deformation_graph.h
@@ -7,7 +7,6 @@
 
 #include <gtsam/geometry/Point3.h>
 #include <gtsam/geometry/Pose3.h>
-#include <gtsam/nonlinear/NonlinearFactor.h>
 #include <gtsam/nonlinear/NonlinearFactorGraph.h>
 #include <gtsam/nonlinear/Values.h>
 #include <pcl/PolygonMesh.h>
@@ -28,15 +27,6 @@
 
 namespace kimera_pgmo {
 
-#if GTSAM_VERSION_MAJOR <= 4 && GTSAM_VERSION_MINOR < 3
-using GtsamJacobianType = boost::optional<gtsam::Matrix&>;
-#define JACOBIAN_DEFAULT \
-  {}
-#else
-using GtsamJacobianType = gtsam::OptionalMatrixType;
-#define JACOBIAN_DEFAULT nullptr
-#endif
-
 /*
  * @brief Struct containing necessary info to add valence factor for a node
  */
@@ -50,75 +40,6 @@ struct NodeValenceInfo {
 using NodeValenceInfoList = std::vector<NodeValenceInfo>;
 
 using EdgeTypeVarianceMap = std::map<pose_graph_tools::PoseGraphEdge::Type, double>;
-
-/*! \brief Define a factor type for edges between two mesh vertices or between a
- * mesh vertex and a pose graph node to be added to deformation graph. Inherited
- * from GTSAM NoiseModelFactor2
- */
-class DeformationEdgeFactor
-    : public gtsam::NoiseModelFactor2<gtsam::Pose3, gtsam::Pose3> {
- private:
-  gtsam::Point3 measurement_;
-
- public:
-  DeformationEdgeFactor(gtsam::Key node1_key,
-                        gtsam::Key node2_key,
-                        const gtsam::Point3& measurement,
-                        gtsam::SharedNoiseModel model)
-      : gtsam::NoiseModelFactor2<gtsam::Pose3, gtsam::Pose3>(
-            model, node1_key, node2_key),
-        measurement_(measurement) {}
-
-  DeformationEdgeFactor(gtsam::Key node1_key,
-                        gtsam::Key node2_key,
-                        const gtsam::Pose3& node1_pose,
-                        const gtsam::Point3& node2_point,
-                        gtsam::SharedNoiseModel model)
-      : gtsam::NoiseModelFactor2<gtsam::Pose3, gtsam::Pose3>(
-            model, node1_key, node2_key) {
-    measurement_ =
-        node1_pose.rotation().inverse().rotate(node2_point - node1_pose.translation());
-  }
-
-  virtual ~DeformationEdgeFactor() {}
-
-  gtsam::Vector evaluateError(const gtsam::Pose3& p1,
-                              const gtsam::Pose3& p2,
-                              GtsamJacobianType H1 = JACOBIAN_DEFAULT,
-                              GtsamJacobianType H2 = JACOBIAN_DEFAULT) const override {
-    // position of node 2 in frame of node 1
-    gtsam::Matrix H_R1, H_t1, H_t2;
-    gtsam::Rot3 R1 = p1.rotation();
-    gtsam::Point3 t1 = p1.translation(H_t1);
-    // New position of node 2 according to deformation p1 of node 1
-    gtsam::Point3 t2_1 = t1 + R1.rotate(measurement_, H_R1);
-    gtsam::Point3 t2_2 = p2.translation(H_t2);
-
-    // Calculate Jacobians
-    if (H1) {
-      Eigen::MatrixXd Jacobian_1 = Eigen::MatrixXd::Zero(3, 6);
-      Jacobian_1.block<3, 3>(0, 0) = H_R1;
-      Jacobian_1 = Jacobian_1 + H_t1;
-      *H1 = Jacobian_1;
-    }
-
-    if (H2) {
-      Eigen::MatrixXd Jacobian_2 = Eigen::MatrixXd::Zero(3, 6);
-      Jacobian_2 = Jacobian_2 - H_t2;
-      *H2 = Jacobian_2;
-    }
-
-    return t2_1 - t2_2;
-  }
-
-  inline gtsam::Point3 measurement() const { return measurement_; }
-
-  gtsam::NonlinearFactor::shared_ptr clone() const override {
-    return gtsam::NonlinearFactor::shared_ptr(new DeformationEdgeFactor(*this));
-  }
-};
-
-#undef JACOBIAN_DEFAULT
 
 class DeformationGraph {
  public:
