+++ kimera_pgmo/include/kimera_pgmo/mesh_frontend_interface.h
@@ -14,6 +14,7 @@
 #include "kimera_pgmo/compression/mesh_compression.h"
 #include "kimera_pgmo/hashing.h"
 #include "kimera_pgmo/utils/common_structs.h"
+#include "kimera_pgmo/utils/graph.h"
 
 namespace kimera_pgmo {
