+++ kimera_pgmo/src/deformation_graph_io.cpp
@@ -11,6 +11,7 @@
 #include <stdexcept>
 
 #include "kimera_pgmo/deformation_graph.h"
+#include "kimera_pgmo/deformation_edge_factor.h"
 #include "kimera_pgmo/utils/common_functions.h"
 
 namespace kimera_pgmo {
