+++ kimera_pgmo/test/suites/compression/test_delta_compression.cpp
@@ -8,6 +8,7 @@
 #include <gtest/gtest.h>
 
 #include <chrono>
+#include <numeric>
 #include <sstream>
 #include <string>
 
@@ -159,10 +160,10 @@ bool faceInFaces(const Face& face, const Faces& faces) {
 
 Faces facesFromDelta(const MeshDelta& delta) {
   Faces faces;
-  faces.insert(faces.end(),
-               delta.face_archive_updates().begin(),
-               delta.face_archive_updates().end());
-  faces.insert(faces.end(), delta.face_updates().begin(), delta.face_updates().end());
+  const auto& archived = delta.archived_faces();
+  faces.insert(faces.end(), archived.begin(), archived.end());
+  const auto& active = delta.faces();
+  faces.insert(faces.end(), active.begin(), active.end());
   return faces;
 }
 
@@ -212,6 +213,7 @@ struct ExpectedDelta {
   void checkTriangles(const Faces& result, const std::vector<size_t>& remapping) const;
 
   void checkMesh(const MeshOffsetInfo& offsets,
+                 const std::vector<traits::Vertex>& prev_vertices,
                  const std::vector<traits::Vertex>& vertices,
                  const std::vector<traits::Face>& faces) const;
 };
@@ -295,11 +297,40 @@ void ExpectedDelta::checkTriangles(const Faces& result,
 }
 
 void ExpectedDelta::checkMesh(const MeshOffsetInfo& offsets,
-                              const std::vector<traits::Vertex>& /* vertices */,
+                              const std::vector<traits::Vertex>& prev_vertices,
+                              const std::vector<traits::Vertex>& vertices,
                               const std::vector<traits::Face>& /* faces */) const {
   EXPECT_EQ(state.archived_vertices, offsets.archived_vertices);
   EXPECT_EQ(state.archived_faces, offsets.archived_faces);
-  // TODO(nathan) think about checking vertices
+
+  std::list<size_t> prev_indices(state.prev_active_vertices);
+  std::iota(prev_indices.begin(), prev_indices.end(), state.archived_vertices);
+
+  MeshOffsetInfo::RemapStats stats;
+  offsets.remapVertexIndices(prev_indices, &stats);
+
+  auto iter = prev_indices.begin();
+  std::map<size_t, size_t> result_remap;
+  for (size_t i = 0; i < state.prev_active_vertices; ++i) {
+    if (stats.deleted_indices.count(i)) {
+      continue;
+    }
+
+    ASSERT_NE(iter, prev_indices.end());
+    result_remap[i] = *iter;
+    ++iter;
+  }
+
+  for (const auto& [prev, curr] : result_remap) {
+    ASSERT_LT(prev, prev_vertices.size());
+    ASSERT_LT(curr, vertices.size());
+    const auto p_prev = prev_vertices.at(prev).pos;
+    const auto p_curr = vertices.at(curr).pos;
+    Eigen::IOFormat fmt(3, Eigen::DontAlignCols, ", ", "; ", "", "", "[", "]");
+    EXPECT_NEAR((p_prev - p_curr).norm(), 0.0, 1.0e-6)
+        << "prev: " << p_prev.format(fmt) << ", curr: " << p_curr.format(fmt)
+        << ", remap: " << prev << " -> " << curr;
+  }
 }
 
 namespace {
@@ -390,151 +421,209 @@ CompressionTestConfiguration test_configurations[] = {
      1.0e-3,
      {
          {{std::nullopt, 100s, {block1_v1}},
-          {{6, 0, 0, 0, 0},  // b1 has 6 vertices
-           {{3, 4, 5}, {6, 7, 8}},
-           {3, 4, 5, 3, 4, 5, 6, 7, 8}}},
+          {
+              {6, 0, 0, 0, 0},  // b1 has 6 vertices
+              {{3, 4, 5}, {6, 7, 8}},
+              {3, 4, 5, 3, 4, 5, 6, 7, 8},
+          }},
          {{101s, 102s, {block1_empty}},
-          {{6, 6, 2, 6, 2},  // replacing previous 6 with archived 6
-           {{3, 4, 5}, {6, 7, 8}},
-           {}}},
+          {
+              {6, 6, 2, 6, 2},  // replacing previous 6 with archived 6
+              {{3, 4, 5}, {6, 7, 8}},
+              {},
+          }},
          {{std::nullopt, 103s, {block1_v1}},
-          {{6, 0, 0, 6, 2},  // new 6 vertices + archival
-           {{12, 13, 14}, {15, 16, 17}},
-           {12, 13, 14, 12, 13, 14, 15, 16, 17}}},
+          {
+              {6, 0, 0, 6, 2},  // new 6 vertices + archival
+              {{12, 13, 14}, {15, 16, 17}},
+              {12, 13, 14, 12, 13, 14, 15, 16, 17},
+          }},
      }},
     {"MultiBlockClearing",
      1.0e-3,
      {
          {{std::nullopt, 100s, {block1_v1, block2_v1}},
-          {{9, 0, 0, 0, 0},  // 9 unique vertices (and 3 faces)
-           {{3, 4, 5}, {12, 13, 14}, {9, 10, 11}},
-           {3, 4, 5, 3, 4, 5, 12, 13, 14, 9, 10, 11, 12, 13, 14}}},
+          {
+              {9, 0, 0, 0, 0},  // 9 unique vertices (and 3 faces)
+              {{3, 4, 5}, {12, 13, 14}, {9, 10, 11}},
+              {3, 4, 5, 3, 4, 5, 12, 13, 14, 9, 10, 11, 12, 13, 14},
+          }},
          {{std::nullopt, 101s, {block1_empty, block2_v1}},
-          {{6, 9, 3, 0, 0},  // 6 unique vertices, clear 9/3 previous
-           {{15, 16, 17}, {18, 19, 20}},
-           {15, 16, 17, 18, 19, 20}}},
+          {
+              {6, 9, 3, 0, 0},  // 6 unique vertices, clear 9/3 previous
+              {{15, 16, 17}, {18, 19, 20}},
+              {15, 16, 17, 18, 19, 20},
+          }},
          {{std::nullopt, 102s, {block1_v1, block2_v1}},
-          {{9, 6, 2, 0, 0},  // 9/3 unique, clear 6/2 previous
-           {{24, 25, 26}, {33, 34, 35}, {30, 31, 32}},
-           {24, 25, 26, 24, 25, 26, 33, 34, 35, 30, 31, 32, 33, 34, 35}}},
+          {
+              {9, 6, 2, 0, 0},  // 9/3 unique, clear 6/2 previous
+              {{24, 25, 26}, {33, 34, 35}, {30, 31, 32}},
+              {24, 25, 26, 24, 25, 26, 33, 34, 35, 30, 31, 32, 33, 34, 35},
+          }},
          {{std::nullopt, 103s, {block1_v1, block2_empty}},
-          {{6, 9, 3, 0, 0},  // 6/2 unique, clear 9/3 previous
-           {{39, 40, 41}, {42, 43, 44}},
-           {39, 40, 41, 39, 40, 41, 42, 43, 44}}},
+          {
+              {6, 9, 3, 0, 0},  // 6/2 unique, clear 9/3 previous
+              {{39, 40, 41}, {42, 43, 44}},
+              {39, 40, 41, 39, 40, 41, 42, 43, 44},
+          }},
      }},
     {"MultiBlockPrune",
      1.0e-3,
      {
          {{std::nullopt, 100s, {block1_v1}},
-          {{6, 0, 0, 0, 0},  // 6/2 unique
-           {{3, 4, 5}, {6, 7, 8}},
-           {3, 4, 5, 3, 4, 5, 6, 7, 8}}},
+          {
+              {6, 0, 0, 0, 0},  // 6/2 unique
+              {{3, 4, 5}, {6, 7, 8}},
+              {3, 4, 5, 3, 4, 5, 6, 7, 8},
+          }},
          {{std::nullopt, 102s, {block2_v1}},
-          {{9, 6, 2, 0, 0},  // 9/3 unique, remove 6/2
-           {{3, 4, 5}, {12, 13, 14}, {9, 10, 11}},
-           {9, 10, 11, 12, 13, 14}}},
+          {
+              {9, 6, 2, 0, 0},  // 9/3 unique, remove 6/2
+              {{3, 4, 5}, {12, 13, 14}, {9, 10, 11}},
+              {9, 10, 11, 12, 13, 14},
+          }},
          {{101s, 103s, {block1_empty, block2_empty}},
-          {{6, 9, 3, 3, 1},  //  6/2 unique, remove 9/3
-           {{3, 4, 5}, {12, 13, 14}},
-           {}}},
+          {
+              {6, 9, 3, 3, 1},  //  6/2 unique, remove 9/3
+              {{3, 4, 5}, {12, 13, 14}},
+              {},
+          }},
          {{std::nullopt, 104s, {block1_v1}},
-          {{9, 3, 0, 3, 1},  // 6/2 unique, archive 3/1, 3/1 pending
-           {{18, 19, 20}, {21, 22, 23}},
-           {18, 19, 20, 18, 19, 20, 21, 22, 23}}},
+          {
+              {9, 3, 0, 3, 1},  // 6/2 unique, archive 3/1, 3/1 pending
+              {{18, 19, 20}, {21, 22, 23}},
+              {18, 19, 20, 18, 19, 20, 21, 22, 23},
+          }},
          {{std::nullopt, 105s, {block1_v1}},
-          {{9, 9, 2, 3, 1},  // 6/2 unique, archive 3/1, 3/1 pending
-           {{27, 28, 29}, {30, 31, 32}},
-           {27, 28, 29, 27, 28, 29, 30, 31, 32}}},
+          {
+              {9, 9, 2, 3, 1},  // 6/2 unique, archive 3/1, 3/1 pending
+              {{27, 28, 29}, {30, 31, 32}},
+              {27, 28, 29, 27, 28, 29, 30, 31, 32},
+          }},
      }},
     {"MultiBlockPartialUpdates",
      1.0e-3,
      {
          {{std::nullopt, 100s, {block1_v1}},
-          {{6, 0, 0, 0, 0},  // 6/2 unique
-           {{3, 4, 5}, {6, 7, 8}},
-           {3, 4, 5, 3, 4, 5, 6, 7, 8}}},
+          {
+              {6, 0, 0, 0, 0},  // 6/2 unique
+              {{3, 4, 5}, {6, 7, 8}},
+              {3, 4, 5, 3, 4, 5, 6, 7, 8},
+          }},
          {{std::nullopt, 102s, {block2_v1}},
-          {{9, 6, 2, 0, 0},  // 9/3 unique, remove 6/2
-           {{3, 4, 5}, {12, 13, 14}, {9, 10, 11}},
-           {9, 10, 11, 12, 13, 14}}},
+          {
+              {9, 6, 2, 0, 0},  // 9/3 unique, remove 6/2
+              {{3, 4, 5}, {12, 13, 14}, {9, 10, 11}},
+              {9, 10, 11, 12, 13, 14},
+          }},
          {{101s, 103s, {block1_empty, block2_v1}},
-          {{9, 9, 3, 3, 1},  // 9/3 unique, remove 9/3
-           {{3, 4, 5}, {18, 19, 20}, {15, 16, 17}, {18, 19, 20}},
-           {15, 16, 17, 18, 19, 20}}},
+          {
+              {9, 9, 3, 3, 1},  // 9/3 unique, remove 9/3
+              {{3, 4, 5}, {18, 19, 20}, {15, 16, 17}, {18, 19, 20}},
+              {15, 16, 17, 18, 19, 20},
+          }},
          {{std::nullopt, 104s, {block2_v1}},
-          {{6, 6, 3, 3, 1},  // 3/2 unique, archive 3/1, pending 3/1
-           {{24, 25, 26}, {21, 22, 23}, {24, 25, 26}},
-           {21, 22, 23, 24, 25, 26}}},
+          {
+              {6, 6, 3, 3, 1},  // 3/2 unique, archive 3/1, pending 3/1
+              {{24, 25, 26}, {21, 22, 23}, {24, 25, 26}},
+              {21, 22, 23, 24, 25, 26},
+          }},
          {{std::nullopt, 105s, {block1_v1, block2_v1}},
-          {{9, 6, 3, 3, 1},  // 6/2 unique, pending 3/1
-           {{39, 40, 41}, {30, 31, 32}, {36, 37, 38}, {39, 40, 41}},
-           {30, 31, 32, 30, 31, 32, 39, 40, 41, 36, 37, 38, 39, 40, 41}}},
+          {
+              {9, 6, 3, 3, 1},  // 6/2 unique, pending 3/1
+              {{39, 40, 41}, {30, 31, 32}, {36, 37, 38}, {39, 40, 41}},
+              {30, 31, 32, 30, 31, 32, 39, 40, 41, 36, 37, 38, 39, 40, 41},
+          }},
          {{std::nullopt, 106s, {block1_empty, block2_empty}},
-          {{3, 9, 4, 3, 1},  // 6/2 unique, pending 3/1
-           {{39, 40, 41}},
-           {}}},
+          {
+              {3, 9, 4, 3, 1},  // 6/2 unique, pending 3/1
+              {{39, 40, 41}},
+              {},
+          }},
      }},
     {"MultiBlockPartialArchive",
      1.0e-3,
      {
          {{std::nullopt, 100s, {block1_v2}},
-          {{6, 0, 0, 0, 0},  // 6/2 unique
-           {{0, 1, 2}, {3, 4, 5}},
-           {0, 1, 2, 3, 4, 5}}},
+          {
+              {6, 0, 0, 0, 0},  // 6/2 unique
+              {{0, 1, 2}, {3, 4, 5}},
+              {0, 1, 2, 3, 4, 5},
+          }},
          {{std::nullopt, 102s, {block2_v2}},
-          {{10, 6, 2, 0, 0},  // 10/4 unique
-           {{0, 1, 2}, {9, 10, 5}, {6, 7, 8}, {9, 10, 11}},
-           {6, 7, 8, 9, 10, 11}}},
+          {
+              {10, 6, 2, 0, 0},  // 10/4 unique
+              {{0, 1, 2}, {9, 10, 5}, {6, 7, 8}, {9, 10, 11}},
+              {6, 7, 8, 9, 10, 11},
+          }},
          {{101s, 103s, {block1_empty, block2_v2}},
-          {{10, 10, 4, 3, 1},  // 10/4 unique, replace all
-           {{0, 1, 2}, {15, 16, 5}, {12, 13, 14}, {15, 16, 17}},
-           {12, 13, 14, 15, 16, 17}}},
+          {
+              {10, 10, 4, 3, 1},  // 10/4 unique, replace all
+              {{0, 1, 2}, {15, 16, 5}, {12, 13, 14}, {15, 16, 17}},
+              {12, 13, 14, 15, 16, 17},
+          }},
          {{std::nullopt, 104s, {block2_empty}},
-          {{3, 7, 3, 3, 1},  // 3/1 pending, remove 7/3
-           {{15, 16, 5}},
-           {}}},
+          {
+              {3, 7, 3, 3, 1},  // 3/1 pending, remove 7/3
+              {{15, 16, 5}},
+              {},
+          }},
      }},
     {"BoundaryVertexRemapping",
      1.0e-3,
      {
          {{std::nullopt, 100s, {block1_v3}},
-          {{8, 0, 0, 0, 0},  // 8/3 unique
-           {{0, 1, 2}, {3, 4, 8}, {6, 7, 8}},
-           {0, 1, 2, 3, 4, 8, 6, 7, 8}}},
+          {
+              {8, 0, 0, 0, 0},  // 8/3 unique
+              {{0, 1, 2}, {3, 4, 8}, {6, 7, 8}},
+              {0, 1, 2, 3, 4, 8, 6, 7, 8},
+          }},
          {{std::nullopt, 102s, {block2_v2}},
-          {{12, 8, 3, 0, 0},  // 12/5 unique, previous 8/3
-           {{0, 1, 2}, {12, 13, 8}, {6, 7, 8}, {9, 10, 11}, {12, 13, 14}},
-           {9, 10, 11, 12, 13, 14}}},
+          {
+              {12, 8, 3, 0, 0},  // 12/5 unique, previous 8/3
+              {{0, 1, 2}, {12, 13, 8}, {6, 7, 8}, {9, 10, 11}, {12, 13, 14}},
+              {9, 10, 11, 12, 13, 14},
+          }},
          {{101s, 103s, {block1_empty, block2_v2}},
-          {{12, 12, 5, 3, 1},  // 12/5 unique, previous 12/5
-           {{0, 1, 2}, {18, 19, 8}, {6, 7, 8}, {15, 16, 17}, {18, 19, 20}},
-           {15, 16, 17, 18, 19, 20}}},
+          {
+              {12, 12, 5, 3, 1},  // 12/5 unique, previous 12/5
+              {{0, 1, 2}, {18, 19, 8}, {6, 7, 8}, {15, 16, 17}, {18, 19, 20}},
+              {15, 16, 17, 18, 19, 20},
+          }},
          {{std::nullopt, 104s, {block2_empty}},
-          {{5, 9, 3, 3, 1},  // 3/1 unique, previous 12/5, archive 3/2
-           {{18, 19, 8}},
-           {}}},
+          {
+              {5, 9, 3, 3, 1},  // 3/1 unique, previous 12/5, archive 3/2
+              {{18, 19, 8}},
+              {},
+          }},
          {{std::nullopt, 105s, {block2_empty}},
-          {{5, 5, 0, 3, 1},  // 3/1 unique, previous 3/1
-           {},
-           {}}},
+          {
+              {5, 5, 0, 3, 1},  // 3/1 unique, previous 3/1
+              {},
+              {},
+          }},
      }},
     {"RepeatedTimestamp",
      1.0e-3,
      {
          {{std::nullopt, 100s, {block1_v1}},
-          {{6, 0, 0, 0, 0},  // 6/2 unique
-           {{3, 4, 5}, {6, 7, 8}},
-           {3, 4, 5, 3, 4, 5, 6, 7, 8}}},
+          {
+              {6, 0, 0, 0, 0},  // 6/2 unique
+              {{3, 4, 5}, {6, 7, 8}},
+              {3, 4, 5, 3, 4, 5, 6, 7, 8},
+          }},
          {{std::nullopt, 100s, {block2_v1}},
-          {{9, 6, 2, 0, 0},  // 9/3 unique, 6/2 prev
-           {{3, 4, 5}, {9, 10, 11}, {12, 13, 14}},
-           {9, 10, 11, 12, 13, 14}}},
+          {
+              {9, 6, 2, 0, 0},  // 9/3 unique, 6/2 prev
+              {{3, 4, 5}, {9, 10, 11}, {12, 13, 14}},
+              {9, 10, 11, 12, 13, 14},
+          }},
      }},
 };
 
 }  // namespace
 
-TEST(TestDeltaCompression, vertexInfoCorrect) {
+TEST(DeltaCompression, vertexInfoCorrect) {
   // base info should have a ref count of 0
   VertexInfo info;
   EXPECT_TRUE(info.notObserved());
@@ -573,7 +662,7 @@ struct DeltaCompressionFixture
     : public testing::TestWithParam<CompressionTestConfiguration> {};
 
 INSTANTIATE_TEST_SUITE_P(
-    TestDeltaCompression,
+    DeltaCompression,
     DeltaCompressionFixture,
     testing::ValuesIn(test_configurations),
     [](const testing::TestParamInfo<DeltaCompressionFixture::ParamType>& info) {
@@ -590,6 +679,7 @@ TEST_P(DeltaCompressionFixture, CompressionCorrect) {
   MeshOffsetInfo offsets;
   std::vector<traits::Face> result_faces;
   std::vector<traits::Vertex> result_vertices;
+  std::vector<traits::Vertex> prev_vertices;
   for (const auto& [input, expected] : config.inputs) {
     if (input.prune_time_ns) {
       compression.archiveBlocksByTime(input.prune_time_ns->count());
@@ -606,8 +696,9 @@ TEST_P(DeltaCompressionFixture, CompressionCorrect) {
     const auto result_indices = flattenRemapping(input.blocks, remapping);
     expected.checkOutput(*output, result_indices);
 
+    prev_vertices = result_vertices;
     output->updateMesh(result_vertices, result_faces, offsets);
-    expected.checkMesh(offsets, result_vertices, result_faces);
+    expected.checkMesh(offsets, prev_vertices, result_vertices, result_faces);
   }
 }
