+++ kimera_pgmo/test/suites/test_mesh_deformation.cpp
@@ -4,9 +4,9 @@
  * @author Yun Chang
  */
 
+#include <gtest/gtest.h>
 #include <gtsam/geometry/Pose3.h>
 
-#include "gtest/gtest.h"
 #include "kimera_pgmo/mesh_deformation.h"
 #include "kimera_pgmo/pcl_mesh_traits.h"
 #include "test_config.h"
