+++ kimera_pgmo/test/suites/utils/test_range_generator.cpp
@@ -2,25 +2,11 @@
 
 #include <numeric>
 
-#include "kimera_pgmo/mesh_traits.h"
-#include "kimera_pgmo/pcl_mesh_traits.h"
 #include "kimera_pgmo/utils/range_generator.h"
 
 namespace kimera_pgmo {
 
-TEST(TestTraits, stampTraitsCorrect) {
-  static_assert(!traits::has_get_stamp<pcl::PointCloud<pcl::PointXYZ>>::value,
-                "pcl::PointXYZ get failed");
-  static_assert(!traits::has_get_stamp<pcl::PointCloud<pcl::PointXYZRGBA>>::value,
-                "pcl::PointXYZRBA get failed");
-  static_assert(traits::has_get_stamp<ConstStampedCloud<pcl::PointXYZ>>::value,
-                "const stamped pcl::PointXYZ failed");
-  static_assert(traits::has_get_stamp<ConstStampedCloud<pcl::PointXYZRGBA>>::value,
-                "const stamped pcl::PointXYZRBA failed");
-  SUCCEED();
-}
-
-TEST(TestTraits, generator) {
+TEST(RangeGenerator, generatorCorrect) {
   RangeGenerator generator(10);
   std::vector<int64_t> result(generator.begin(), generator.end());
   std::vector<int64_t> expected(10);
